+++ browser/js/funcs.js
@@ -34,7 +34,7 @@ function getUsernames (chat_, shouldTruncate) {
 
 function isCurrentChat (chat_) {
   if (window.currentChatId === DUMMY_CHAT_ID) {
-    return  !chatsHash[chat_.id];
+    return !window.chatListHash[chat_.id];
   } else {
     return chat_.id === window.currentChatId;
   }
@@ -42,7 +42,9 @@ function isCurrentChat (chat_) {
 
 function setActive (el) {
   let currentActive = document.querySelector('.chat-list ul li.active');
-  if (currentActive) currentActive.classList.remove('active');
+  if (currentActive) {
+    currentActive.classList.remove('active');
+  }
   el.classList.add('active');
 
   // close opened emoji pane
@@ -129,8 +131,12 @@ function addSubmitHandler (chat_) {
 }
 
 function sendAttachment(filePath, chat_) {
+  // @todo: pass this as argument instead
+  window.notifiedChatId = chat_.id
+  notify('Your file is being uploaded', true)
+
   var recipients = chat_.accounts.map((account) => account.id)
-  ipcRenderer.send('upload', { filePath, recipients, isNewChat: !chat_.id })
+  ipcRenderer.send('upload', { filePath, recipients, isNewChat: !chat_.id, chatId: chat_.id })
 }
 
 function addAttachmentSender(chat_) {
@@ -139,6 +145,7 @@ function addAttachmentSender(chat_) {
     fileInput.click();
     fileInput.onchange = () => {
       sendAttachment(fileInput.files[0].path, chat_);
+      fileInput.value = '';
     }
   }
 }
@@ -148,8 +155,8 @@ function addNotification (el, chat_) {
     return
   }
 
-  const isNew = (chatsHash && chatsHash[chat_.id] &&
-    chatsHash[chat_.id].items[0].id !== chat_.items[0].id
+  const isNew = (window.chatListHash[chat_.id] &&
+    window.chatListHash[chat_.id].items[0].id !== chat_.items[0].id
   );
   if (isNew) unreadChats[chat_.id] = chat_;
 
@@ -158,6 +165,7 @@ function addNotification (el, chat_) {
       markAsRead(chat_.id, el);
     } else {
       el.classList.add('notification');
+      // @todo pass this as an argument instead
       window.notifiedChatId = el.getAttribute("id");
       if (isNew && window.shouldNotify && !window.isWindowFocused) {
         notify(`new message from ${getUsernames(chat_)}`);
@@ -166,8 +174,10 @@ function addNotification (el, chat_) {
   }
 }
 
-function notify (message) {
-  ipcRenderer.send('increase-badge-count');
+function notify(message, noBadgeCountIncrease) {
+  if (!noBadgeCountIncrease) {
+    ipcRenderer.send('increase-badge-count');
+  }
   const notification = new Notification('IG:dm Desktop', {
     body: message
   });